# web-service-gin/docs/swagger.yaml
basePath: /
definitions:
  main.album:
    properties:
      artist:
        type: string
      id:
        type: integer
      price:
        type: number
      title:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a generated swagger API from the Album API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Carlo's Album API
  version: "1.0"
paths:
  /album:
    get:
      consumes:
      - application/json
      description: gets one (1) album using a query.
      parameters:
      - description: album search by id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched album.
          schema:
            $ref: '#/definitions/main.album'
      summary: gets one (1) album using a query.
      tags:
      - albums
  /albums:
    get:
      consumes:
      - application/json
      description: gets all albums from the database.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the albums
          schema:
            items:
              $ref: '#/definitions/main.album'
            type: array
      summary: gets all albums
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: postAlbums adds a new album from JSON received in the request body.
      parameters:
      - description: Album object
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/main.album'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully added album.
          schema:
            $ref: '#/definitions/main.album'
      summary: adds a new album
      tags:
      - albums
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"